test(mirrord-browser): remove instrumentation-related test additions
