+++ src/__tests__/useHeaderRules.test.ts
@@ -1,5 +1,12 @@
 import { renderHook, act } from '@testing-library/react';
-import { armCanary, cancelCanary } from '../headerObservation';
+
+// Mock analytics module
+const mockCapture = jest.fn();
+jest.mock('../analytics', () => ({
+    capture: (...args: unknown[]) => mockCapture(...args),
+    emitUserBlocked: jest.fn(),
+    emitUserSucceeded: jest.fn(),
+}));
 
 jest.mock('../headerObservation', () => {
     const actual = jest.requireActual('../headerObservation');
@@ -10,16 +17,6 @@ jest.mock('../headerObservation', () => {
         notifyHeaderObserved: jest.fn(),
     };
 });
-
-// Mock analytics module
-const mockCapture = jest.fn();
-const mockEmitUserBlocked = jest.fn();
-const mockEmitUserSucceeded = jest.fn();
-jest.mock('../analytics', () => ({
-    capture: (...args: unknown[]) => mockCapture(...args),
-    emitUserBlocked: (...args: unknown[]) => mockEmitUserBlocked(...args),
-    emitUserSucceeded: (...args: unknown[]) => mockEmitUserSucceeded(...args),
-}));
 
 // Mock chrome API
 const mockGetDynamicRules = jest.fn();
@@ -72,8 +69,6 @@ import { useHeaderRules } from '../hooks/useHeaderRules';
 describe('useHeaderRules analytics', () => {
     beforeEach(() => {
         jest.clearAllMocks();
-        mockEmitUserBlocked.mockClear();
-        mockEmitUserSucceeded.mockClear();
         (
             chrome.runtime as { lastError: chrome.runtime.LastError | null }
         ).lastError = null;
@@ -590,418 +585,6 @@ describe('useHeaderRules analytics', () => {
         });
     });
 
-    describe('emitUserBlocked typed reasons', () => {
-        it('emits header_rule_save_failed on activate failure', async () => {
-            mockGetDynamicRules.mockImplementation(
-                (cb: (rules: chrome.declarativeNetRequest.Rule[]) => void) =>
-                    cb([])
-            );
-            mockStorageGet.mockImplementation(
-                (
-                    _keys: string[],
-                    cb: (result: Record<string, unknown>) => void
-                ) =>
-                    cb({
-                        defaults: {
-                            headerName: 'X-Test',
-                            headerValue: 'val',
-                        },
-                    })
-            );
-            mockUpdateDynamicRules.mockImplementation(
-                (_opts: unknown, cb: () => void) => {
-                    (
-                        chrome.runtime as {
-                            lastError: chrome.runtime.LastError | null;
-                        }
-                    ).lastError = { message: 'activate failed' };
-                    cb();
-                    (
-                        chrome.runtime as {
-                            lastError: chrome.runtime.LastError | null;
-                        }
-                    ).lastError = null;
-                }
-            );
-
-            const { result } = renderHook(() => useHeaderRules());
-            await act(async () => {});
-
-            await act(async () => {
-                result.current.handleToggle();
-            });
-
-            expect(mockEmitUserBlocked).toHaveBeenCalledWith(
-                'header_rule_save_failed',
-                'user_action',
-                expect.objectContaining({ error: 'activate failed' })
-            );
-        });
-
-        it('emits header_rule_remove_failed on remove failure', async () => {
-            const activeRule: chrome.declarativeNetRequest.Rule = {
-                id: 1,
-                priority: 1,
-                action: {
-                    type: 'modifyHeaders' as chrome.declarativeNetRequest.RuleActionType,
-                    requestHeaders: [
-                        {
-                            header: 'X-Test',
-                            operation:
-                                'set' as chrome.declarativeNetRequest.HeaderOperation,
-                            value: 'val',
-                        },
-                    ],
-                },
-                condition: { urlFilter: '|' },
-            };
-            mockGetDynamicRules.mockImplementation(
-                (cb: (rules: chrome.declarativeNetRequest.Rule[]) => void) =>
-                    cb([activeRule])
-            );
-            mockUpdateDynamicRules.mockImplementation(
-                (_opts: unknown, cb: () => void) => {
-                    (
-                        chrome.runtime as {
-                            lastError: chrome.runtime.LastError | null;
-                        }
-                    ).lastError = { message: 'remove failed' };
-                    cb();
-                    (
-                        chrome.runtime as {
-                            lastError: chrome.runtime.LastError | null;
-                        }
-                    ).lastError = null;
-                }
-            );
-
-            const { result } = renderHook(() => useHeaderRules());
-            await act(async () => {});
-
-            await act(async () => {
-                result.current.handleToggle();
-            });
-
-            expect(mockEmitUserBlocked).toHaveBeenCalledWith(
-                'header_rule_remove_failed',
-                'user_action',
-                expect.objectContaining({ error: 'remove failed' })
-            );
-        });
-
-        it('emits header_rule_save_failed on save update_rules failure', async () => {
-            const activeRule: chrome.declarativeNetRequest.Rule = {
-                id: 1,
-                priority: 1,
-                action: {
-                    type: 'modifyHeaders' as chrome.declarativeNetRequest.RuleActionType,
-                    requestHeaders: [
-                        {
-                            header: 'X-Test',
-                            operation:
-                                'set' as chrome.declarativeNetRequest.HeaderOperation,
-                            value: 'val',
-                        },
-                    ],
-                },
-                condition: { urlFilter: '|' },
-            };
-            mockGetDynamicRules.mockImplementation(
-                (cb: (rules: chrome.declarativeNetRequest.Rule[]) => void) =>
-                    cb([activeRule])
-            );
-            mockUpdateDynamicRules.mockImplementation(
-                (_opts: unknown, cb: () => void) => {
-                    (
-                        chrome.runtime as {
-                            lastError: chrome.runtime.LastError | null;
-                        }
-                    ).lastError = { message: 'dnr save failed' };
-                    cb();
-                    (
-                        chrome.runtime as {
-                            lastError: chrome.runtime.LastError | null;
-                        }
-                    ).lastError = null;
-                }
-            );
-
-            const { result } = renderHook(() => useHeaderRules());
-            await act(async () => {});
-
-            act(() => {
-                result.current.setHeaderName('X-New');
-                result.current.setHeaderValue('new');
-            });
-
-            await act(async () => {
-                result.current.handleSave();
-            });
-
-            expect(mockEmitUserBlocked).toHaveBeenCalledWith(
-                'header_rule_save_failed',
-                'user_action',
-                expect.objectContaining({ error: 'dnr save failed' })
-            );
-        });
-
-        it('emits header_rule_save_failed on save storage_write failure', async () => {
-            mockStorageSet.mockImplementation(
-                (_data: unknown, cb: () => void) => {
-                    (
-                        chrome.runtime as {
-                            lastError: chrome.runtime.LastError | null;
-                        }
-                    ).lastError = { message: 'storage write failed' };
-                    cb();
-                    (
-                        chrome.runtime as {
-                            lastError: chrome.runtime.LastError | null;
-                        }
-                    ).lastError = null;
-                }
-            );
-
-            const { result } = renderHook(() => useHeaderRules());
-
-            act(() => {
-                result.current.setHeaderName('X-Test');
-                result.current.setHeaderValue('value');
-            });
-
-            await act(async () => {
-                result.current.handleSave();
-            });
-
-            expect(mockEmitUserBlocked).toHaveBeenCalledWith(
-                'header_rule_save_failed',
-                'user_action',
-                expect.objectContaining({ error: 'storage write failed' })
-            );
-        });
-
-        it('emits header_rule_reset_failed on reset storage_remove failure', async () => {
-            mockStorageGet.mockImplementation(
-                (
-                    _keys: string[],
-                    cb: (result: Record<string, unknown>) => void
-                ) =>
-                    cb({
-                        defaults: {
-                            headerName: 'X-Default',
-                            headerValue: 'defaultval',
-                        },
-                    })
-            );
-            mockStorageRemove.mockImplementation(
-                (_keys: string[], cb: () => void) => {
-                    (
-                        chrome.runtime as {
-                            lastError: chrome.runtime.LastError | null;
-                        }
-                    ).lastError = { message: 'storage remove failed' };
-                    cb();
-                    (
-                        chrome.runtime as {
-                            lastError: chrome.runtime.LastError | null;
-                        }
-                    ).lastError = null;
-                }
-            );
-
-            const { result } = renderHook(() => useHeaderRules());
-
-            await act(async () => {
-                result.current.handleReset();
-            });
-
-            expect(mockEmitUserBlocked).toHaveBeenCalledWith(
-                'header_rule_reset_failed',
-                'user_action',
-                expect.objectContaining({ error: 'storage remove failed' })
-            );
-        });
-
-        it('emits header_rule_reset_failed on reset update_rules failure', async () => {
-            mockStorageGet.mockImplementation(
-                (
-                    _keys: string[],
-                    cb: (result: Record<string, unknown>) => void
-                ) =>
-                    cb({
-                        defaults: {
-                            headerName: 'X-Default',
-                            headerValue: 'defaultval',
-                        },
-                    })
-            );
-            mockStorageRemove.mockImplementation(
-                (_keys: string[], cb: () => void) => cb()
-            );
-            mockUpdateDynamicRules.mockImplementation(
-                (_opts: unknown, cb: () => void) => {
-                    (
-                        chrome.runtime as {
-                            lastError: chrome.runtime.LastError | null;
-                        }
-                    ).lastError = { message: 'dnr reset failed' };
-                    cb();
-                    (
-                        chrome.runtime as {
-                            lastError: chrome.runtime.LastError | null;
-                        }
-                    ).lastError = null;
-                }
-            );
-
-            const { result } = renderHook(() => useHeaderRules());
-
-            await act(async () => {
-                result.current.handleReset();
-            });
-
-            expect(mockEmitUserBlocked).toHaveBeenCalledWith(
-                'header_rule_reset_failed',
-                'user_action',
-                expect.objectContaining({ error: 'dnr reset failed' })
-            );
-        });
-    });
-
-    describe('emitUserSucceeded typed reasons', () => {
-        it('emits header_rule_activated on successful activate', async () => {
-            mockGetDynamicRules.mockImplementation(
-                (cb: (rules: chrome.declarativeNetRequest.Rule[]) => void) =>
-                    cb([])
-            );
-            mockStorageGet.mockImplementation(
-                (
-                    _keys: string[],
-                    cb: (result: Record<string, unknown>) => void
-                ) =>
-                    cb({
-                        defaults: {
-                            headerName: 'X-Test',
-                            headerValue: 'val',
-                        },
-                    })
-            );
-            mockUpdateDynamicRules.mockImplementation(
-                (_opts: unknown, cb: () => void) => cb()
-            );
-            mockStorageRemove.mockImplementation(
-                (_keys: string[], cb: () => void) => cb()
-            );
-
-            const { result } = renderHook(() => useHeaderRules());
-            await act(async () => {});
-
-            await act(async () => {
-                result.current.handleToggle();
-            });
-
-            expect(mockEmitUserSucceeded).toHaveBeenCalledWith(
-                'header_rule_activated',
-                'user_action'
-            );
-        });
-
-        it('emits header_rule_removed on successful remove', async () => {
-            const activeRule: chrome.declarativeNetRequest.Rule = {
-                id: 1,
-                priority: 1,
-                action: {
-                    type: 'modifyHeaders' as chrome.declarativeNetRequest.RuleActionType,
-                    requestHeaders: [
-                        {
-                            header: 'X-Test',
-                            operation:
-                                'set' as chrome.declarativeNetRequest.HeaderOperation,
-                            value: 'val',
-                        },
-                    ],
-                },
-                condition: { urlFilter: '|' },
-            };
-            mockGetDynamicRules.mockImplementation(
-                (cb: (rules: chrome.declarativeNetRequest.Rule[]) => void) =>
-                    cb([activeRule])
-            );
-            mockUpdateDynamicRules.mockImplementation(
-                (_opts: unknown, cb: () => void) => cb()
-            );
-            mockStorageRemove.mockImplementation(
-                (_keys: string[], cb: () => void) => cb()
-            );
-
-            const { result } = renderHook(() => useHeaderRules());
-            await act(async () => {});
-
-            await act(async () => {
-                result.current.handleToggle();
-            });
-
-            expect(mockEmitUserSucceeded).toHaveBeenCalledWith(
-                'header_rule_removed',
-                'user_action'
-            );
-        });
-
-        it('emits header_rule_saved on successful save', async () => {
-            mockStorageSet.mockImplementation(
-                (_data: unknown, cb: () => void) => cb()
-            );
-
-            const { result } = renderHook(() => useHeaderRules());
-
-            act(() => {
-                result.current.setHeaderName('X-Test');
-                result.current.setHeaderValue('value');
-            });
-
-            await act(async () => {
-                result.current.handleSave();
-            });
-
-            expect(mockEmitUserSucceeded).toHaveBeenCalledWith(
-                'header_rule_saved',
-                'user_action'
-            );
-        });
-
-        it('emits header_rule_reset on successful reset', async () => {
-            mockStorageGet.mockImplementation(
-                (
-                    _keys: string[],
-                    cb: (result: Record<string, unknown>) => void
-                ) =>
-                    cb({
-                        defaults: {
-                            headerName: 'X-Default',
-                            headerValue: 'defaultval',
-                        },
-                    })
-            );
-            mockStorageRemove.mockImplementation(
-                (_keys: string[], cb: () => void) => cb()
-            );
-            mockUpdateDynamicRules.mockImplementation(
-                (_opts: unknown, cb: () => void) => cb()
-            );
-
-            const { result } = renderHook(() => useHeaderRules());
-
-            await act(async () => {
-                result.current.handleReset();
-            });
-
-            expect(mockEmitUserSucceeded).toHaveBeenCalledWith(
-                'header_rule_reset',
-                'user_action'
-            );
-        });
-    });
-
     describe('handleShare', () => {
         beforeEach(() => {
             mockClipboardWriteText.mockClear();
@@ -1083,102 +666,6 @@ describe('useHeaderRules analytics', () => {
             });
 
             expect(mockClipboardWriteText).not.toHaveBeenCalled();
-        });
-    });
-
-    describe('canary wiring', () => {
-        beforeEach(() => {
-            (armCanary as jest.Mock).mockClear();
-            (cancelCanary as jest.Mock).mockClear();
-        });
-
-        it('arms canary on successful save with flow=header_injector', async () => {
-            mockStorageSet.mockImplementation(
-                (_data: unknown, cb: () => void) => cb()
-            );
-
-            const { result } = renderHook(() => useHeaderRules());
-
-            act(() => {
-                result.current.setHeaderName('X-Canary');
-                result.current.setHeaderValue('canary-val');
-            });
-
-            await act(async () => {
-                result.current.handleSave();
-            });
-
-            expect(armCanary).toHaveBeenCalledWith({
-                headerName: 'X-Canary',
-                flow: 'header_injector',
-            });
-        });
-
-        it('cancels canary on successful remove', async () => {
-            const activeRule: chrome.declarativeNetRequest.Rule = {
-                id: 1,
-                priority: 1,
-                action: {
-                    type: 'modifyHeaders' as chrome.declarativeNetRequest.RuleActionType,
-                    requestHeaders: [
-                        {
-                            header: 'X-Test',
-                            operation:
-                                'set' as chrome.declarativeNetRequest.HeaderOperation,
-                            value: 'val',
-                        },
-                    ],
-                },
-                condition: { urlFilter: '|' },
-            };
-            mockGetDynamicRules.mockImplementation(
-                (cb: (rules: chrome.declarativeNetRequest.Rule[]) => void) =>
-                    cb([activeRule])
-            );
-            mockUpdateDynamicRules.mockImplementation(
-                (_opts: unknown, cb: () => void) => cb()
-            );
-            mockStorageRemove.mockImplementation(
-                (_keys: string[], cb: () => void) => cb()
-            );
-
-            const { result } = renderHook(() => useHeaderRules());
-            await act(async () => {});
-
-            await act(async () => {
-                result.current.handleToggle();
-            });
-
-            expect(cancelCanary).toHaveBeenCalled();
-        });
-
-        it('cancels canary on successful reset', async () => {
-            mockStorageGet.mockImplementation(
-                (
-                    _keys: string[],
-                    cb: (result: Record<string, unknown>) => void
-                ) =>
-                    cb({
-                        defaults: {
-                            headerName: 'X-Default',
-                            headerValue: 'defaultval',
-                        },
-                    })
-            );
-            mockStorageRemove.mockImplementation(
-                (_keys: string[], cb: () => void) => cb()
-            );
-            mockUpdateDynamicRules.mockImplementation(
-                (_opts: unknown, cb: () => void) => cb()
-            );
-
-            const { result } = renderHook(() => useHeaderRules());
-
-            await act(async () => {
-                result.current.handleReset();
-            });
-
-            expect(cancelCanary).toHaveBeenCalled();
         });
     });
